+++ src/Api/Setting/Option.php
@@ -17,6 +17,7 @@ use JOOOSI_FON;
 use JooosiFon\Api\AbstractApi;
 use JooosiFon\Api\ApiInterface;
 use JooosiFon\Api\Support\RequestValidator;
+use JooosiFon\Upgrade\LegacyRebrandUpgrade;
 use WP_REST_Request;
 use WP_REST_Response;
 use WP_REST_Server;
@@ -107,9 +108,32 @@ class Option extends AbstractApi implements ApiInterface
             ], 500);
         }
 
+        if ($this->shouldDeleteLegacyData($options)
+            && ! (new LegacyRebrandUpgrade())->deleteLegacyData()) {
+            return new WP_REST_Response([
+                'message' => __('The legacy Yabe Webfont data cleanup was incomplete. Any remaining data was preserved.', 'jooosi-fon'),
+            ], 500);
+        }
+
         do_action('f!jooosi/fon/api/setting/option:after_store', $options);
         do_action_deprecated('f!yabe/webfont/api/setting/option:after_store', [$options], '2.1.0', 'f!jooosi/fon/api/setting/option:after_store');
 
         return $this->index($wprestRequest);
+    }
+
+    /**
+     * @param mixed $options
+     */
+    private function shouldDeleteLegacyData($options): bool
+    {
+        if (is_object($options)) {
+            $options = get_object_vars($options);
+        }
+
+        if (! is_array($options) || ! is_array($options['misc'] ?? null)) {
+            return false;
+        }
+
+        return ! empty($options['misc']['delete_legacy_data']);
     }
 }